upload codes

+++ BOJ/1008.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 float main()
 {
-    float a,b;
-    scanf("%f %f",&a,&b);
+    double a,b;
+    scanf("%lf %lf",&a,&b);
     printf("%.9f",a/b);
     return 0;
 }
